:tada: Initial commit with webview & server implementation
The first commit to the project with a basic project structure ready. The NanoServer file contains a small http server implementation to serve the static files present in the project.

There is a basic callback template present in the CustomCallback file which is used to implement JavaScript callbacks within Java.