:card_file_box: Add database handler for operations users and settings
Created a DatabaseHandler java file to act as a high-level wrapper between the internal logic and database operations. Removed org.reflections as a dependency and created a custom Reflections file to handle runtime class searching.

This commit also fixes a bug where dark/light mode wouldn't be synced on using the back button.